// src/app/interfaces/proposta.ts
export interface Proposta {
    qsc?: string;
    central?: string;
    especifica?: string;
    probl?: string;
    hipoteses?: string;
    cts?: string;
    conteudo?: string;
    cooperacao?: string;
    planejamento?: string;
    investigacao?: string;
    dados?: string;
    retomada?: string;
    acao?: string;
    comunicacao?: string;
    redes?: string;
    uid?: string;
    compartilhar?: boolean;
    ativo?: string;
}